+++ packages/preact_signals/lib/preact_signals.dart
@@ -10,6 +10,8 @@ export 'src/signals.dart'
         Signal,
         MutableSignal,
         ReadonlySignal;
-export 'src/extensions.dart';
+export 'src/extensions/stream.dart';
+export 'src/extensions/future.dart';
 export 'src/signal_state.dart';
 export 'src/future_signal.dart';
+export 'src/stream_signal.dart';