+++ packages/preact_signals/lib/src/extensions.dart
@@ -1,2 +1,0 @@
-export 'extensions/future.dart';
-export 'extensions/stream.dart';