+++ packages/preact_signals/lib/src/future_signal.dart
@@ -10,8 +10,11 @@ import 'package:preact_signals/preact_signals.dart';
 ///
 /// If success then the result will be [SignalValue]
 class FutureSignal<T> extends Signal<SignalState> {
+  /// Future [Duration] to wait before timing out
+  final Duration? timeout;
+
   /// Creates a [FutureSignal] that wraps a [Future]
-  FutureSignal(this._getFuture) : super(SignalLoading()) {
+  FutureSignal(this._getFuture, {this.timeout}) : super(SignalLoading()) {
     _init();
   }
 
@@ -26,11 +29,20 @@ class FutureSignal<T> extends Signal<SignalState> {
     if (peek() is! SignalLoading) {
       value = SignalLoading();
     }
-
-    _getFuture().then((value) {
+    var f = _getFuture();
+    if (timeout != null) {
+      f = f.timeout(timeout!, onTimeout: () {
+        throw SignalTimeout();
+      });
+    }
+    f.then((value) {
       this.value = SignalValue(value);
     }).catchError((error) {
-      value = SignalError(error);
+      if (error is SignalTimeout) {
+        value = error;
+      } else {
+        value = SignalError(error);
+      }
     });
   }
 }